1) Add a possibility for using a full matrix format for serial iterative solvers:
Use Full Matrix for Radiation = Logical

Potential use for view factor normalization and computing of gebhardt factors and
radiosities.

2) Do radiosity computation with only one task and distribute the result to the rest
of the participating tasks. This avoids the memory bandwidth problems when doing
multiple linear system solutions at the same time, at least for non-supercomputer
systems. The new scheme is several times faster on my laptop when running in parallel
with ~4-6 tasks).