Use mGetElementDOFs in the generation of generic mortar constraints
The SUBROUTINE AddProjectorWeakGeneric is revised so that the indices for
the basis functions are obtained by calling the library utility mGetElementDOFs.
The consistent indexing is thus obtained from one place and need not be recreated
in connection with mortar constraints. Some other subroutines related to mortar
conditions could also be revised in the same way to ensure consistency and 
to simplify the maintenance of the code.