Functional changes related to mortar keywords
This change resolves some keyword conflicts of the earlier code and it may also
change how the commands of a sif file are now interpreted.

The keyword Level Projector Generic has not anymore impact on whether the mortar
constraints are generated weakly or not. For this purpose use only the keywords
Galerkin Projector or Level Projector Strong. If this information is not provided,
the constraints are now generated in a strong manner corresponding to
the choices Galerkin Projector = False and/or Level Projector Strong = True.
For a more fine-tuned selection the keyword Level Projector Nodes Strong is
also available.

On the other hand the keyword Level Projector Generic = True can be used to
select the newest subroutines for generating mortar constraints. This has effects
on the generation of both weak and strong constraints.